+++ doc/releases.md
@@ -255,12 +255,11 @@ We continue to clean up Param's API ([#734](https://github.com/holoviz/param/pul
     - `ClassSelector(class_=(str, int))`: OK
     - It's possible that in the future the signature of these two parameters will be aligned with the other parameters to accept `default` as first and only positional argument, but for now please use an explicit keyword so that your code will be compatible with all versions.
 - Parameter slots:
-  - `List._class`: use instead `item_type`.
+  - `List.class_`: use instead `item_type`.
   - `Number.set_hook`: no replacement
 - `param.__init__` module:
   - `param.produce_value`: no replacement
   - `param.as_unicode`: no replacement
-  - `param.is_ordered_dict`: no replacement
   - `param.is_ordered_dict`: no replacement
   - `param.hashable`: no replacement
   - `param.named_objs`: no replacement